\name{calcule.poids.Chen}
\alias{calcule.poids.Chen}
\title{
function to compute weights
}
\description{
blabla
}
\usage{
calcule.poids.Chen(xl,y,ind.par,rep.par,alpha,lc,klc)
}
\arguments{
  \item{xl}{
}
  \item{y}{
}
  \item{ind.par}{
}
  \item{rep.par}{
}
  \item{alpha}{
}
  \item{lc}{
}
  \item{klc}{
}
}
\value{
w
}
\author{
Alexandre Bureau and Jordie Croteau
}


