\name{fat2Lpoly-internal}

%% Documented as internal :
%% <LIST>
\alias{alleles2sums}
\alias{cov.score.interfunction}
\alias{cov.score.poly}
\alias{scores.covs}
\alias{design.polytomous}
\alias{ibd.terms}
\alias{ibd.terms.w}
\alias{score.poly}
\alias{produits.paires}
\alias{converti.terme}
\alias{calcule.poids.Chen}
\alias{calcule.poids.alphafixe}
%% </LIST>

\title{Internal fat2Lpoly Functions}
\description{
  Internal fat2Lpoly functions
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
