\name{fat2Lpoly.allSNPs}
\alias{fat2Lpoly.allSNPs}
\docType{data}
\title{
sample output of the function \code{fat2Lpoly.withinR}
}
\description{
This example output of the function \code{fat2Lpoly.withinR} is provided to test the function \code{get.scores.pvalues} by executing the example code in the \code{get.scores.pvalues} documentation.
}
\usage{data(fat2Lpoly.allSNPs)}
\format{
  The format is:
List of 2
 $ scores.covs.all.SNPs:List of 5
  ..$ :List of 2
  .. ..$ scores.mat: num [1:100, 1:5] 0.5 0 0 0 0.25 ...
  .. ..$ cov.mat   : num [1:100, 1:5, 1:5] 0.128 0 0 0 0.159 ...
  ..$ :List of 2
  .. ..$ scores.mat: num [1:100, 1:5] 0.5 0 0 0 0.25 ...
  .. ..$ cov.mat   : num [1:100, 1:5, 1:5] 0.128 0 0 0 0.159 ...
  ..$ :List of 2
  .. ..$ scores.mat: num [1:100, 1:5] 0.5 0 0 0 0.25 ...
  .. ..$ cov.mat   : num [1:100, 1:5, 1:5] 0.128 0 0 0 0.159 ...
  ..$ :List of 2
  .. ..$ scores.mat: num [1:100, 1:5] 0.5 0 0 0 0.25 ...
  .. ..$ cov.mat   : num [1:100, 1:5, 1:5] 0.128 0 0 0 0.159 ...
  ..$ :List of 2
  .. ..$ scores.mat: num [1:100, 1:5] 0.5 0 0 0 0.25 ...
  .. ..$ cov.mat   : num [1:100, 1:5, 1:5] 0.128 0 0 0 0.159 ...
 $ snp.names.mat       : chr [1:5, 1:2] "snp4.loc1" "snp4.loc1" "snp4.loc1" "snp4.loc1" ...
}
\examples{
data(fat2Lpoly.allSNPs)
}
\keyword{datasets}
