\name{ped.x.all}
\alias{ped.x.all}
\docType{data}
\title{
sample output of the function \code{read.merlin.files}
}
\description{
This example output of the function \code{read.merlin.files} is provided to test the function \code{fat2Lpoly.withinR} by executing the example code in the \code{fat2Lpoly.withinR} documentation.
}
\usage{data(ped.x.all)}
\format{
  The format is:
List of 7
 $ ped         :'data.frame':	1600 obs. of  4 variables:
  ..$ fam.id     : int [1:1600] 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 ...
  ..$ subject.ids: int [1:1600] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ y1         : num [1:1600] 1 1 1 1 1 2 1 1 2 1 ...
  ..$ y2         : num [1:1600] 1 1 1 1 1 1 1 1 2 1 ...
 $ x.all       : num [1:1600, 1:16] 0.5 0.5 1 0 1 0.5 0 0 0.5 0.5 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:16] "snp1.loc1" "snp2.loc1" "snp3.loc1" "snp4.loc1" ...
 $ MA.table    :'data.frame':	16 obs. of  2 variables:
  ..$ snp.names.dat: Factor w/ 16 levels "snp1.loc1","snp1.loc2",..: 1 3 5 7 9 11 13 15 2 4 ...
  ..$ MA           : int [1:16] 1 1 1 2 1 1 1 1 1 1 ...
 $ ibd.dat.list:List of 2
  ..$ :'data.frame':	108800 obs. of  7 variables:
  .. ..$ FAMILY: int [1:108800] 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 ...
  .. ..$ ID1   : int [1:108800] 1 2 2 7 7 7 8 8 8 8 ...
  .. ..$ ID2   : int [1:108800] 1 1 2 1 2 7 1 2 7 8 ...
  .. ..$ MARKER: chr [1:108800] "1_1_mrk:" "1_1_mrk:" "1_1_mrk:" "1_1_mrk:" ...
  .. ..$ P0    : num [1:108800] 0 1 0 1 1 0 1 1 1 0 ...
  .. ..$ P1    : num [1:108800] 0 0 0 0 0 0 0 0 0 0 ...
  .. ..$ P2    : num [1:108800] 1 0 1 0 0 1 0 0 0 1 ...
  ..$ :'data.frame':	108800 obs. of  7 variables:
  .. ..$ FAMILY: int [1:108800] 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 1000051 ...
  .. ..$ ID1   : int [1:108800] 1 2 2 7 7 7 8 8 8 8 ...
  .. ..$ ID2   : int [1:108800] 1 1 2 1 2 7 1 2 7 8 ...
  .. ..$ MARKER: chr [1:108800] "2_1_mrk:" "2_1_mrk:" "2_1_mrk:" "2_1_mrk:" ...
  .. ..$ P0    : num [1:108800] 0 1 0 1 1 0 1 1 1 0 ...
  .. ..$ P1    : num [1:108800] 0 0 0 0 0 0 0 0 0 0 ...
  .. ..$ P2    : num [1:108800] 1 0 1 0 0 1 0 0 0 1 ...
 $ y1.name     : chr "endo"
 $ y2.name     : chr "pheno"
 $ ibdfilenames: chr [1:2] "loc1.ibd" "loc2.ibd"
}
\examples{
data(ped.x.all)
}
\keyword{datasets}
